requirements.txt
PyYaml
joblib
fastavro
confluent-kafka
schedule
scikit-learn
apscheduler
jinja2